refactor(sprint/ai-musician/1): render-video route 분리 + seed 배포 가드 (RF-1, RF-2)
- render-video/route.ts (492줄) → lib/videoRender/{abort,fonts,canvas,ffmpeg,media}.ts + 161줄 오케스트레이터로 분리
- /api/seed: NODE_ENV=production 조기 반환 추가 — 배포 환경에서 private/ 파일 접근 차단
